requirements.txt
fastapi[standard]
pydantic
python-decouple
docker
paramiko
langchain
langchain-ollama
proxmoxer